requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anthropic==0.53.0
    # via polyglot-mcp (pyproject.toml)
anyio==4.9.0
    # via
    #   anthropic
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
click==8.2.1
    # via uvicorn
distro==1.9.0
    # via anthropic
fastapi==0.115.12
    # via polyglot-mcp (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   mcp
httpx-sse==0.4.0
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
jiter==0.10.0
    # via anthropic
mcp==1.9.3
    # via polyglot-mcp (pyproject.toml)
pydantic==2.11.5
    # via
    #   polyglot-mcp (pyproject.toml)
    #   anthropic
    #   fastapi
    #   mcp
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via mcp
python-dotenv==1.1.0
    # via
    #   polyglot-mcp (pyproject.toml)
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
sse-starlette==2.3.6
    # via mcp
starlette==0.46.2
    # via
    #   fastapi
    #   mcp
typing-extensions==4.14.0
    # via
    #   anthropic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.34.3
    # via
    #   polyglot-mcp (pyproject.toml)
    #   mcp
